Relocate webinar list to a separate component
